requirements.txt
asgiref==3.5.2
Django==4.0.4
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.2
pytz==2022.1
sqlparse==0.4.2
factory-boy==3.2.1
Faker==13.15.0
drf-yasg==1.20.0
django-filter==22.1
